// frontend/src/types/websocket.types.ts
export type AuKey =
  | "AU1"
  | "AU2"
  | "AU4"
  | "AU5"
  | "AU6"
  | "AU7"
  | "AU9"
  | "AU10"
  | "AU12"
  | "AU15"
  | "AU20"
  | "AU23"
  | "AU25"
  | "AU26";